+++ src/main/java/dev/jasmine/carrybabyanimals/reunion/ParentReunionMatch.java
@@ -1,0 +1,6 @@
+package dev.jasmine.carrybabyanimals.reunion;
+
+import net.minecraft.world.entity.animal.Animal;
+
+public record ParentReunionMatch(Animal baby, Animal adult, String babyFeedbackName) {
+}